LLMs-Code/Gemini/requirements.txt
pandas==2.1.0
google-generativeai==0.1.0
